refactor: 각 역할별로 Room service 분리
